Refactor to allow graphUtils.c to be graph.c
